requirements.txt
python-dotenv>=1.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
pdf2image>=1.16.0
pytesseract>=0.3.10
openai>=1.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
langgraph>=0.2.0
langchain-core>=0.2.0
fastapi>=0.100.0
pydantic>=2.0.0
python-multipart>=0.0.6
uvicorn>=0.23.0
streamlit>=1.28.0
requests>=2.31.0